Add: xah-fly-keys-search function to conditionally use swoop/swiper (#5)
There is a bit of an unusual situation for keybinding on the top level
command mode map. It is dynamically rebuilt by a function every time command
mode is toggled on, so the key gets rebound in a hook after the xah-fly-key-map
has been populated. This solution probably will not fly for those using
non-qwerty keyboards, and maybe should get a PR to the xah-fly-keys package
instead. There is precedence with the M-x function using smex, helm-M-x or
standard execute-extended-command.